+++ node/clients/catalog.ts
@@ -1,4 +1,10 @@
-import { AppClient, InstanceOptions, IOContext, RequestConfig, SegmentData } from '@vtex/api'
+import {
+  AppClient,
+  InstanceOptions,
+  IOContext,
+  RequestConfig,
+  SegmentData,
+} from '@vtex/api'
 import { stringify } from 'qs'
 
 import { CatalogCrossSellingTypes } from '../resolvers/catalog/utils'
@@ -8,8 +14,13 @@ interface AutocompleteArgs {
   searchTerm: string
 }
 
-const inflightKey = ({baseURL, url, params, headers}: RequestConfig) => {
-  return baseURL! + url! + stringify(params, {arrayFormat: 'repeat', addQueryPrefix: true}) + `&segmentToken=${headers['x-vtex-segment']}`
+const inflightKey = ({ baseURL, url, params, headers }: RequestConfig) => {
+  return (
+    baseURL! +
+    url! +
+    stringify(params, { arrayFormat: 'repeat', addQueryPrefix: true }) +
+    `&segmentToken=${headers['x-vtex-segment']}`
+  )
 }
 
 /** Catalog API
@@ -20,100 +31,114 @@ export class Catalog extends AppClient {
     super('vtex.catalog-api-proxy', ctx, opts)
   }
 
-  public product = (slug: string) => this.get<Product[]>(
-    `/pub/products/search/${slug && slug.toLowerCase()}/p`,
-    {metric: 'catalog-product'}
-  )
+  public product = (slug: string) =>
+    this.get<Product[]>(
+      `/pub/products/search/${slug && slug.toLowerCase()}/p`,
+      { metric: 'catalog-product' }
+    )
 
-  public productByEan = (id: string) => this.get<Product[]>(
-    `/pub/products/search?fq=alternateIds_Ean:${id}`,
-    {metric: 'catalog-productByEan'}
-  )
+  public productByEan = (id: string) =>
+    this.get<Product[]>(`/pub/products/search?fq=alternateIds_Ean:${id}`, {
+      metric: 'catalog-productByEan',
+    })
 
-  public productsByEan = (ids: string[]) => this.get<Product[]>(
-    `/pub/products/search?${ids.map(id => `fq=alternateIds_Ean:${id}`).join('&')}`,
-    {metric: 'catalog-productByEan'}
-  )
+  public productsByEan = (ids: string[]) =>
+    this.get<Product[]>(
+      `/pub/products/search?${ids
+        .map(id => `fq=alternateIds_Ean:${id}`)
+        .join('&')}`,
+      { metric: 'catalog-productByEan' }
+    )
 
-  public productById = (id: string) => this.get<Product[]>(
-    `/pub/products/search?fq=productId:${id}`,
-    {metric: 'catalog-productById'}
-  )
+  public productById = (id: string) =>
+    this.get<Product[]>(`/pub/products/search?fq=productId:${id}`, {
+      metric: 'catalog-productById',
+    })
 
-  public productsById = (ids: string[]) => this.get<Product[]>(
-    `/pub/products/search?${ids.map(id => `fq=productId:${id}`).join('&')}`,
-    {metric: 'catalog-productById'}
-  )
+  public productsById = (ids: string[]) =>
+    this.get<Product[]>(
+      `/pub/products/search?${ids.map(id => `fq=productId:${id}`).join('&')}`,
+      { metric: 'catalog-productById' }
+    )
 
-  public productByReference = (id: string) => this.get<Product[]>(
-    `/pub/products/search?fq=alternateIds_RefId:${id}`,
-    {metric: 'catalog-productByReference'}
-  )
-  
-  public productsByReference = (ids: string[]) => this.get<Product[]>(
-    `/pub/products/search?${ids.map(id => `fq=alternateIds_RefId:${id}`).join('&')}`,
-    {metric: 'catalog-productByReference'}
-  )
+  public productByReference = (id: string) =>
+    this.get<Product[]>(`/pub/products/search?fq=alternateIds_RefId:${id}`, {
+      metric: 'catalog-productByReference',
+    })
 
-  public productBySku = (skuIds: string[]) => this.get<Product[]>(
-    `/pub/products/search?${skuIds.map(skuId => `fq=skuId:${skuId}`).join('&')}`,
-    {metric: 'catalog-productBySku'}
-  )
+  public productsByReference = (ids: string[]) =>
+    this.get<Product[]>(
+      `/pub/products/search?${ids
+        .map(id => `fq=alternateIds_RefId:${id}`)
+        .join('&')}`,
+      { metric: 'catalog-productByReference' }
+    )
+
+  public productBySku = (skuIds: string[]) =>
+    this.get<Product[]>(
+      `/pub/products/search?${skuIds
+        .map(skuId => `fq=skuId:${skuId}`)
+        .join('&')}`,
+      { metric: 'catalog-productBySku' }
+    )
 
   public products = (args: SearchArgs, useRaw = false) => {
     const method = useRaw ? this.getRaw : this.get
-    return method<Product[]>(
-      this.productSearchUrl(args),
-      {metric: 'catalog-products'}
-    )
+    return method<Product[]>(this.productSearchUrl(args), {
+      metric: 'catalog-products',
+    })
   }
 
   public productsQuantity = async (args: SearchArgs) => {
-    const {headers: {resources}} = await this.getRaw(this.productSearchUrl(args))
+    const {
+      headers: { resources },
+    } = await this.getRaw(this.productSearchUrl(args))
     const quantity = resources.split('/')[1]
     return parseInt(quantity, 10)
   }
 
-  public brands = () => this.get<Brand[]>(
-    '/pub/brand/list',
-    {metric: 'catalog-brands'}
-  )
+  public brands = () =>
+    this.get<Brand[]>('/pub/brand/list', { metric: 'catalog-brands' })
 
-  public categories = (treeLevel: number) => this.get<Category[]>(
-    `/pub/category/tree/${treeLevel}/`,
-    {metric: 'catalog-categories'}
-  )
+  public categories = (treeLevel: number) =>
+    this.get<Category[]>(`/pub/category/tree/${treeLevel}/`, {
+      metric: 'catalog-categories',
+    })
 
   public facets = (facets: string = '') => {
     const [path, options] = decodeURI(facets).split('?')
     return this.get(
-      `/pub/facets/search/${encodeURI(`${path.trim()}${options ? '?' + options : ''}`)}`,
-      {metric: 'catalog-facets'}
+      `/pub/facets/search/${encodeURI(
+        `${path.trim()}${options ? '?' + options : ''}`
+      )}`,
+      { metric: 'catalog-facets' }
     )
   }
 
-  public category = (id: string | number) => this.get(
-    `/pub/category/${id}`,
-    {metric: 'catalog-category'}
-  )
+  public category = (id: string | number) =>
+    this.get(`/pub/category/${id}`, { metric: 'catalog-category' })
 
-  public crossSelling = (id: string, type: CatalogCrossSellingTypes) => this.get<Product[]>(
-    `/pub/products/crossselling/${type}/${id}`,
-    {metric: 'catalog-crossSelling'}
-  )
+  public crossSelling = (id: string, type: CatalogCrossSellingTypes) =>
+    this.get<Product[]>(`/pub/products/crossselling/${type}/${id}`, {
+      metric: 'catalog-crossSelling',
+    })
 
-  public autocomplete = ({maxRows, searchTerm}: AutocompleteArgs) => this.get<{itemsReturned: Item[]}>(
-    `/buscaautocomplete?maxRows=${maxRows}&productNameContains=${encodeURIComponent(searchTerm)}`,
-    {metric: 'catalog-autocomplete'}
-  )
+  public autocomplete = ({ maxRows, searchTerm }: AutocompleteArgs) =>
+    this.get<{ itemsReturned: Item[] }>(
+      `/buscaautocomplete?maxRows=${maxRows}&productNameContains=${encodeURIComponent(
+        searchTerm
+      )}`,
+      { metric: 'catalog-autocomplete' }
+    )
 
   private get = <T = any>(url: string, config: RequestConfig = {}) => {
-    const segmentData: SegmentData | undefined = (this.context! as CustomIOContext).segment
+    const segmentData: SegmentData | undefined = (this
+      .context! as CustomIOContext).segment
     const { channel: salesChannel = '' } = segmentData || {}
 
     config.params = {
       ...config.params,
-      ...!!salesChannel && {sc: salesChannel},
+      ...(!!salesChannel && { sc: salesChannel }),
     }
 
     config.inflightKey = inflightKey
@@ -122,12 +147,13 @@ export class Catalog extends AppClient {
   }
 
   private getRaw = <T = any>(url: string, config: RequestConfig = {}) => {
-    const segmentData: SegmentData | undefined = (this.context! as CustomIOContext).segment
+    const segmentData: SegmentData | undefined = (this
+      .context! as CustomIOContext).segment
     const { channel: salesChannel = '' } = segmentData || {}
 
     config.params = {
       ...config.params,
-      ...!!salesChannel && {sc: salesChannel},
+      ...(!!salesChannel && { sc: salesChannel }),
     }
 
     config.inflightKey = inflightKey
@@ -146,14 +172,23 @@ export class Catalog extends AppClient {
     to = 9,
     map = '',
     hideUnavailableItems = false,
+    filledFields = '',
   }: SearchArgs) => {
     const sanitizedQuery = encodeURIComponent(decodeURIComponent(query).trim())
     if (hideUnavailableItems) {
       const segmentData = (this.context as CustomIOContext).segment
-      salesChannel = segmentData && segmentData.channel.toString() || ''
+      salesChannel = (segmentData && segmentData.channel.toString()) || ''
     }
-    return (
-      `/pub/products/search/${sanitizedQuery}?${category && !query && `&fq=C:/${category}/`}${(specificationFilters && specificationFilters.length > 0 && specificationFilters.map(filter => `&fq=${filter}`)) || ''}${priceRange && `&fq=P:[${priceRange}]`}${collection && `&fq=productClusterIds:${collection}`}${salesChannel && `&fq=isAvailablePerSalesChannel_${salesChannel}:1`}${orderBy && `&O=${orderBy}`}${map && `&map=${map}`}${from > -1 && `&_from=${from}`}${to > -1 && `&_to=${to}`}`
-    )
+    return `/pub/products/search/${sanitizedQuery}?${category &&
+      !query &&
+      `&fq=C:/${category}/`}${filledFields &&
+      `&fq=filledFields:${filledFields}`}${(specificationFilters &&
+      specificationFilters.length > 0 &&
+      specificationFilters.map(filter => `&fq=${filter}`)) ||
+      ''}${priceRange && `&fq=P:[${priceRange}]`}${collection &&
+      `&fq=productClusterIds:${collection}`}${salesChannel &&
+      `&fq=isAvailablePerSalesChannel_${salesChannel}:1`}${orderBy &&
+      `&O=${orderBy}`}${map && `&map=${map}`}${from > -1 &&
+      `&_from=${from}`}${to > -1 && `&_to=${to}`}`
   }
 }